proc composition `compose` now delegates to Proc#<< supported from Ruby 2.6
Ruby 2.6 Proc#>> implentation does not concern passed object is callable.

So the following code chains `>>` with Symbol raises error on Ruby 2.6
Becase of Ruby2.6's Proc#>> does not convert passed symbol
to proc when combine it but lambda_driver does.

```
  arr.map(&:to_s >> :upcase >> :to_sym)
  # =>(undefined method `call' for :to_sym:Symbol)
```

To fix this behaivor, lambda_driver override Proc#>> and Proc#<<